requirements.txt
stripe
pydantic
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asgiref==3.11.0
attrs==25.4.0
black==25.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
curlify==3.0.0
Django==5.2.8
django-admin-sortable2==2.3
django-appconf==1.2.0
django-cors-headers==4.9.0
django-countries==8.2.0
django-debug-toolbar==6.1.0
django-filter==25.2
django-imagekit==6.0.0
django-js-asset==3.1.2
django-mptt==0.18.0
django-webpack-loader==3.2.2
djangorestframework==3.16.1
drf-standardized-errors==0.15.0
facebook_business==24.0.1
frozenlist==1.8.0
gunicorn==23.0.0
idna==3.11
Markdown==3.10
multidict==6.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pilkit==3.0
pillow==12.0.0
platformdirs==4.5.0
propcache==0.4.1
psycopg==3.2.13
psycopg-binary==3.2.13
pycountry==24.6.1
python-decouple==3.8
pytokens==0.3.0
requests==2.32.5
ruff==0.14.7
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
urllib3==2.6.1
yarl==1.22.0
